backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
Pillow==10.1.0
boto3==1.29.7
stripe==7.8.0
hypothesis==6.92.1
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
bcrypt==4.1.2
